Revoke selected batch or single credential
New dialog for viewing and revoking issued credentials for the loaded
DID. The dialog lets you select either a batch or a single credential
for revocation. Future release should support multiple selection both
when issuing credentials and when revoking. This is not supported with
this release.

The functionality now makes use of an incomplete and non-functioning
implementation that saves batch info to file system. The selected
credential is revoked successfully using the Prism node, but the
information like operation hash is not fully implemented for saving to
the file system. This is incomplete because we will not use time
developing file functionality that will be removed in next release. In
the next release, we save the information to a database instead of file
system metadata files.