Vault implemented as Derby database
When dids and contacts are created, they are now saved in a database and
read from the database instead of from a xml file structure.
Functionality still contain hard-coding of for example password, but
this will fixed in a future release when the databases are encrypted and
passwords will have to be correct in order to see the database contents.