Credential issuing and settings
This commit was big. It contains code for saving and retrieving
information from database. No more use of xml files in solution.

Another big addition to this commit is the settings object that is
shared throughout the application. The settings object contains a
Session object which is specific for the currently selected DID. When a
DID is selected as current, the session part of settings is cleared and
reinitialised according to current DID.