Partially functioning drag and drop
Not all cases are currently supported or working, but Drop Source and
Drop Targets, including listeners are now part of the solution and
responding to drag and drop actions.

Still to do, is to get all different kinds of drag and drops to work as
expected. Will be easier to debug when we have added right-click menu
that supports adding and removing elements of different types