Basic right-click menu with two actions
Added right-click context menu to credential builder tree viewer.
Supported actions at this point is "Add Object" and "Delete".
- "Delete" deletes the selected tree element, or the complete tree if no
element is selected.
- "Add Object" adds an item as child to the selected item if it is JSON
list or object. If selected item is a JSON property or JSON primitive,
the item is added as sibling to selected element

Issues:
- Added element is currently hard-coded to a random number called new
value. Not able to decide value or name at this point
- No message if element added already exists. This cause overwritten
values at present