Save and load JSON templates and dnd drop on object bug fixed
Save JSON as template button added. No anonymisation of JSON is done,
but a skeleton method was added to ensure this is not forgot

Load JSON template button also added. Reads JSON files from computer.
Will need to be extended in the future for reading from cloud locations

Drag and drop on JSONObjects had a bug which now is fixed